refactor(normal): move handlers into separate file
Also rework the result view
