+++ internal/views/normal2/normal2.go
@@ -1,18 +1,13 @@
 package normal2
 
 import (
-	"fmt"
 	"log/slog"
-	"os"
-	"strings"
 
 	"github.com/charmbracelet/bubbles/help"
-	"github.com/charmbracelet/bubbles/key"
 	"github.com/charmbracelet/bubbles/list"
 	"github.com/charmbracelet/bubbles/textinput"
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
 	"github.com/nobe4/gh-not/internal/actors"
 	"github.com/nobe4/gh-not/internal/config"
 	"github.com/nobe4/gh-not/internal/notifications"
@@ -65,241 +60,29 @@ func Init(n notifications.Notifications, actors actors.ActorsMap, keymap config.
 func (m model) Init() tea.Cmd { return nil }
 
 func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	var cmd tea.Cmd
-
+	slog.Debug("update", "msg", msg)
 	switch msg := msg.(type) {
 
 	case tea.WindowSizeMsg:
-		m.handleResize(msg)
-		return m, nil
+		return m.handleResize(msg)
+
+	case ResultUpdateMsg:
+		m.result.SetContent(msg.content)
 
 	case ApplyCommandMsg:
-		slog.Debug("apply command", "command", msg.Command)
-		actor, ok := m.actors[msg.Command]
-		if !ok {
-			m.result.SetContent(fmt.Sprintf("Invalid command %s\nPress %s to continue ...", msg.Command, m.list.KeyMap.Quit.Keys()))
-			return m, nil
-		}
-
-		m.resultStrings = []string{}
-		m.currentActor = actor
-		m.processQueue = msg.Items
-
-		// TODO: the rendering should be done only in one place
-		if len(m.processQueue) == 0 {
-			m.result.SetContent(fmt.Sprintf("done, press %s to continue ...", m.list.KeyMap.Quit.Keys()))
-		} else {
-			m.result.SetContent(fmt.Sprintf("%d more ...", len(m.processQueue)))
-		}
-		return m, m.applyNext()
+		return msg.apply(m)
 
 	case AppliedCommandMsg:
-		slog.Debug("applied command", "message", msg.Message)
-		m.processQueue = m.processQueue[1:]
-
-		m.resultStrings = append(m.resultStrings, msg.Message)
-
-		// TODO: the rendering should be done only in one place
-		content := lipgloss.JoinVertical(lipgloss.Top, m.resultStrings...)
-		if len(m.processQueue) == 0 {
-			content = lipgloss.JoinVertical(lipgloss.Top, content, fmt.Sprintf("done, press %s to continue ...", m.list.KeyMap.Quit.Keys()))
-		} else {
-			content = lipgloss.JoinVertical(lipgloss.Top, content, fmt.Sprintf("%d more ...", len(m.processQueue)))
-		}
-
-		m.result.SetContent(content)
-
-		return m, m.applyNext()
+		return msg.apply(m)
 
 	case CleanListMsg:
-		// TODO: move into a method
-		items := []list.Item{}
-		for _, e := range m.list.Items() {
-			if i, ok := e.(item); ok {
-				if !i.notification.Meta.Done {
-					items = append(items, e)
-				}
-			}
-		}
-		m.list.SetItems(items)
+		return msg.apply(m)
 
 	case tea.KeyMsg:
-		if m.showResult {
-			return m, m.handleResult(msg)
-		}
-
-		if m.command.Focused() {
-			return m, m.handleCommand(msg)
-		}
-
-		if m.list.FilterState() == list.Filtering {
-			return m, m.handleFiltering(msg)
-		}
-
-		return m, m.handleBrowsing(msg)
+		return m.handleKeyMsg(msg)
 	}
 
+	var cmd tea.Cmd
 	m.list, cmd = m.list.Update(msg)
 	return m, cmd
-}
-
-func (m *model) handleCommand(msg tea.KeyMsg) tea.Cmd {
-	var cmd tea.Cmd
-
-	switch {
-	case key.Matches(msg, m.keymap.CommandAccept):
-		// TODO: move into a method
-		slog.Debug("blur command")
-		command := m.command.Value()
-		m.command.SetValue("")
-		m.command.Blur()
-		m.showResult = true
-		return m.applyCommand(command)
-
-	case key.Matches(msg, m.keymap.CommandCancel):
-		// TODO: move into a method
-		slog.Debug("blur command")
-		m.command.SetValue("")
-		m.command.Blur()
-		return nil
-	}
-
-	m.command, cmd = m.command.Update(msg)
-	return cmd
-}
-
-type ApplyCommandMsg struct {
-	Items   []item
-	Command string
-}
-
-func (m model) applyCommand(command string) tea.Cmd {
-	return func() tea.Msg {
-		selected := []item{}
-
-		for _, i := range m.list.Items() {
-			n, ok := i.(item)
-
-			if !ok {
-				continue
-			}
-			if n.selected {
-				selected = append(selected, n)
-			}
-		}
-
-		return ApplyCommandMsg{Items: selected, Command: command}
-	}
-}
-
-type AppliedCommandMsg struct {
-	Message string
-}
-type CleanListMsg struct {
-	Message string
-}
-
-func (m model) applyNext() tea.Cmd {
-	return func() tea.Msg {
-		if len(m.processQueue) == 0 {
-			slog.Debug("no more command to apply")
-			return CleanListMsg{}
-		}
-
-		current, tail := m.processQueue[0], m.processQueue[1:]
-		m.processQueue = tail
-
-		slog.Debug("apply next", "notification", current.notification.String())
-
-		message := ""
-		out := &strings.Builder{}
-		if err := m.currentActor.Run(current.notification, out); err != nil {
-			message = fmt.Sprintf("Error for '%s': %s", current.notification.Subject.Title, err.Error())
-		} else {
-			message = out.String()
-		}
-
-		return AppliedCommandMsg{Message: message}
-	}
-}
-
-func (m *model) handleResult(msg tea.KeyMsg) tea.Cmd {
-	var cmd tea.Cmd
-
-	switch {
-
-	case key.Matches(msg, m.list.KeyMap.ShowFullHelp):
-		m.help.ShowAll = !m.help.ShowAll
-		// TODO: why is this not showing the full help?
-		slog.Debug("toggle help", "showAll", m.help.ShowAll)
-
-	case key.Matches(msg, m.list.KeyMap.Quit):
-		m.showResult = false
-		return nil
-	}
-
-	m.result, cmd = m.result.Update(msg)
-	return cmd
-}
-
-func (m *model) handleBrowsing(msg tea.KeyMsg) tea.Cmd {
-	slog.Debug("browsing", "key", msg.String())
-
-	switch {
-	case key.Matches(msg, m.list.KeyMap.ShowFullHelp):
-		m.help.ShowAll = !m.help.ShowAll
-		slog.Debug("toggle help", "showAll", m.help.ShowAll)
-
-	case key.Matches(msg, m.keymap.Toggle):
-		if i, ok := m.list.SelectedItem().(item); ok {
-			i.selected = !i.selected
-			slog.Debug("toggle selected", "item", i.notification.Subject.Title, "selected", i.selected)
-
-			return m.list.SetItem(m.list.GlobalIndex(), i)
-		}
-
-	case key.Matches(msg, m.keymap.All):
-		// TODO: move into a method
-		items := []list.Item{}
-		for _, e := range m.list.Items() {
-			if i, ok := e.(item); ok {
-				i.selected = true
-				items = append(items, i)
-			}
-		}
-		m.list.SetItems(items)
-
-	case key.Matches(msg, m.keymap.None):
-		// TODO: move into a method
-		items := []list.Item{}
-		for _, e := range m.list.Items() {
-			if i, ok := e.(item); ok {
-				i.selected = false
-				items = append(items, i)
-			}
-		}
-		m.list.SetItems(items)
-
-	case key.Matches(msg, m.keymap.Open):
-		// TODO: move into a method
-		current, ok := m.list.SelectedItem().(item)
-		if ok {
-			m.actors["open"].Run(current.notification, os.Stderr)
-		}
-
-	case key.Matches(msg, m.keymap.CommandMode):
-		slog.Debug("focus command")
-		m.command.Focus()
-	}
-
-	var cmd tea.Cmd
-	m.list, cmd = m.list.Update(msg)
-	return cmd
-}
-
-func (m *model) handleFiltering(msg tea.KeyMsg) tea.Cmd {
-	slog.Debug("filtering", "key", msg.String())
-	var cmd tea.Cmd
-	m.list, cmd = m.list.Update(msg)
-	return cmd
 }
